pyproject.toml
[project]
name = "building-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.65.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
]
